Update AddCardScreen to use DeckRepository

+++ feature/home/src/main/java/io/kayt/refluent/feature/home/HomeNavigation.kt
@@ -9,7 +9,7 @@ data object HomeRoute
 
 fun NavGraphBuilder.home(
     onAddDeckClick : () -> Unit,
-    onDeckClick : (Int) -> Unit
+    onDeckClick : (Long) -> Unit
 ) {
     composable<HomeRoute> {
         HomeScreen(
